Clase 20/requirements.txt
google-generativeai
openai
python-dotenv
groq
